% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_autocorrelation.R
\name{calculate_autocorrelation}
\alias{calculate_autocorrelation}
\title{Calculate autocorrelation after removal of each observation}
\usage{
calculate_autocorrelation(mat, min_pairs = 20, method = c("pearson",
  "kendall", "spearman"))
}
\arguments{
\item{mat}{a numeric matrix, with nodes (e.g., analytes such as genes, 
proteins, or metabolites) in columns, and samples in rows}

\item{min_pairs}{minimum number of paired, non-missing observations 
to calculate a correlation coefficient; correlations between vectors with 
fewer than this number of paired observations will be replaced with
\code{NA}}

\item{method}{the correlation coefficient to be computed; one of 
\code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}; 
can be abbreviated}
}
\value{
a matrix with identical dimensions to the input matrix, containing
  the autocorrelation for each non-missing observation
}
\description{
Conventionally, autocorrelation refers to the correlation between
a signal and a delayed copy of that signal. 
In \code{modern}, autocorrelation is calculated as the correlation 
between the global interaction profile of a given node, represented 
as a vector of correlation coefficients to all other nodes in the network,
before and after the removal of a given data point.
The autocorrelation is calculated for each data point in turn, and
observations that have a large impact on the global interaction profile are
flagged as outliers.
}
