% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fwer.R
\name{calculate_fwer}
\alias{calculate_fwer}
\title{Calculate the outlier z score corresponding to a given family-wise error rate}
\usage{
calculate_fwer(mat, alpha = 0.05)
}
\arguments{
\item{mat}{a numeric matrix, with nodes (e.g., analytes such as genes, 
proteins, or metabolites) in columns, and samples in rows}

\item{alpha}{type I error associated with the hypothesis test;
defaults to \code{0.05}}
}
\value{
the threshold below which z scores can be considered outliers
}
\description{
Use Bonferroni correction to calculate the Z score that corresponds to a 
family-wide error rate at a given alpha.
}
