% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_z_scores.R
\name{calculate_z_scores}
\alias{calculate_z_scores}
\title{Calculate z scores for outliers}
\usage{
calculate_z_scores(autocor, bins = NA)
}
\arguments{
\item{autocor}{the matrix of autocorrelations}

\item{bins}{optionally, the number of bins into which to group nodes on the
basis of the number of observations}
}
\value{
a matrix with identical dimensions to the input matrix, containing
  the autocorrelation Z score assigned to each non-missing observation
}
\description{
Convert a matrix of autocorrelations to a matrix of 
Z scores, such that the matrix has a mean of zero and a standard deviation
of one. If the matrix contains missing values, this scaling is performed for 
each group of columns with equivalent numbers of missing values separately.
Optionally, if there are many possible numbers of missing values, the z score
can be calculated for approximately equal sized bins of missing value counts
using the \code{bins} parameter.
}
