% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_ESD.R
\name{generalized_ESD}
\alias{generalized_ESD}
\title{Detect outliers using the generalized Extreme Studentized Deviate test}
\usage{
generalized_ESD(x, max_outliers = 5, alpha = 0.05)
}
\arguments{
\item{x}{distribution to find outliers in}

\item{max_outliers}{the prespecified maximum number of outliers; defaults
to \code{5}}

\item{alpha}{type I error associated with the hypothesis test; 
defaults to \code{0.05}}
}
\value{
a vector of the same length as the input, with outliers masked
}
\description{
Identify outliers within a distribution of numeric values using
Rosner's generalized Extreme Studentized Deviate (ESD) test. Unlike the
Grubbs or Tietjen-Moore tests, the generalized ESD test does not require
the number of outliers to be prespecified.
}
\details{
Full details are provided in:

Rosner, Bernard (May 1983), Percentage Points for a Generalized ESD 
Many-Outlier Procedure, \emph{Technometrics}, 25(2), pp. 165-172.
}
